requirements.txt
pygerrit2
python-nmap
requests